+++ banzai/photometry.py
@@ -12,8 +12,19 @@ from banzai.stages import Stage
 from banzai.data import DataTable
 from banzai import logs
 
+from skimage import measure
+
 logger = logging.getLogger('banzai')
 sep.set_sub_object_limit(int(1e6))
+
+
+def radius_of_contour(contour, source):
+    x = contour[:, 1]
+    y = contour[:, 0]
+    x_center = (source['xmax'] - source['xmin'] + 1) / 2.0 - 0.5
+    y_center = (source['ymax'] - source['ymin'] + 1) / 2.0 - 0.5
+
+    return np.percentile(np.sqrt((x - x_center)**2.0 + (y - y_center)** 2.0), 90)
 
 
 class SourceDetector(Stage):
@@ -93,13 +104,6 @@ class SourceDetector(Stage):
                 sources['fluxerr{0}'.format(diameter)] = fluxerr
                 sources['flag'] |= flag
 
-            # Calculate the FWHMs of the stars:
-            fwhm = 2.0 * (np.log(2) * (sources['a'] ** 2.0 + sources['b'] ** 2.0)) ** 0.5
-            sources['fwhm'] = fwhm
-
-            # Cut individual bright pixels. Often cosmic rays
-            sources = sources[fwhm > 1.0]
-
             # Measure the flux profile
             flux_radii, flag = sep.flux_radius(data, sources['x'], sources['y'],
                                                6.0 * sources['a'], [0.25, 0.5, 0.75],
@@ -109,8 +113,22 @@ class SourceDetector(Stage):
             sources['fluxrad50'] = flux_radii[:, 1]
             sources['fluxrad75'] = flux_radii[:, 2]
 
+            # Cut individual bright pixels. Often cosmic rays
+            sources = sources[sources['fluxrad50'] > 0.5]
+
+            # Calculate the FWHMs of the stars:
+            # Here we estimate contours
+            for source in sources[sources['flag'] == 0]:
+                for ratio, keyword in zip([0.5, 0.1], ['fwhm', 'fwtm']):
+                    contours = measure.find_contours(data[source['ymin']: source['ymax'] + 1,
+                                                     source['xmin']: source['xmax'] + 1],
+                                                     ratio * source['peak'])
+                    # If there are multiple contours like a donut might have take the outer
+                    contour_radii = [radius_of_contour(contour, source) for contour in contours]
+                    source[keyword] = 2.0 * np.nanmax(contour_radii)
+
             # Calculate the windowed positions
-            sig = 2.0 / 2.35 * sources['fluxrad50']
+            sig = 2.0 / 2.35 * sources['fwhm']
             xwin, ywin, flag = sep.winpos(data, sources['x'], sources['y'], sig)
             sources['flag'] |= flag
             sources['xwin'] = xwin
@@ -177,6 +195,8 @@ class SourceDetector(Stage):
             catalog['background'].description = 'Average background value in the aperture'
             catalog['fwhm'].unit = 'pixel'
             catalog['fwhm'].description = 'FWHM of the object'
+            catalog['fwtm'].unit = 'pixel'
+            catalog['fwtm'].description = 'Full-Width Tenth Maximum'
             catalog['a'].unit = 'pixel'
             catalog['a'].description = 'Semi-major axis of the object'
             catalog['b'].unit = 'pixel'
@@ -222,11 +242,15 @@ class SourceDetector(Stage):
                 good_objects = np.logical_and(good_objects, np.logical_not(np.isnan(catalog[quantity])))
             if good_objects.sum() == 0:
                 image.meta['L1FWHM'] = ('NaN', '[arcsec] Frame FWHM in arcsec')
+                image.meta['L1FWTM'] = ('NaN', 'Ratio of FWHM to Full-Width Tenth Max')
+
                 image.meta['L1ELLIP'] = ('NaN', 'Mean image ellipticity (1-B/A)')
                 image.meta['L1ELLIPA'] = ('NaN', '[deg] PA of mean image ellipticity')
             else:
                 seeing = np.median(catalog['fwhm'][good_objects]) * image.pixel_scale
                 image.meta['L1FWHM'] = (seeing, '[arcsec] Frame FWHM in arcsec')
+                image.meta['L1FWTM'] = (np.median(catalog['fwtm'][good_objects] / catalog['fwhm'][good_objects]),
+                                        'Ratio of FWHM to Full-Width Tenth Max')
 
                 mean_ellipticity = stats.sigma_clipped_mean(catalog['ellipticity'][good_objects], 3.0)
                 image.meta['L1ELLIP'] = (mean_ellipticity, 'Mean image ellipticity (1-B/A)')